use bits to to efficently calculate joined peices
Consider a plot peice, it can have connections in one of four places,
top, bottom, left, or right.  The blank peice has no connections and the
cross has all four.  Using this idea we can express all plot peices as
numbers representing their configuration, e.g. ├ is 0111 or 7, and ╮ is
1001 or 9.  Now we can combine peices simply by appling bitwise OR!