continue returning buffer contents even if EOF
Before, if an EOF was encountered, 0 was returned and no attempt was
made to parse remaining data in the buffer.  This meant if you had small
input and a huge buffer size, only the first call plot_file_input_read
would return anything, since it would read the entire input at once.